Moved everything around, added some more classes
