+++ Gridcoin.WebApi/Controllers/GridcoinController.cs
