+++ Gridcoin.WebApi/Controllers/OAuthController.cs
