+++ Gridcoin.WebApi/Models/Constants.cs
