+++ Gridcoin.WebApi/Models/GridcoinSettings.cs
