+++ Gridcoin.WebApi/Models/Payment.cs
