+++ Gridcoin.WebApi/Program.cs
