+++ Gridcoin.WebApi/Properties/launchSettings.json
