+++ Gridcoin.WebApi/Services/GridcoinService.cs
@@ -1,7 +1,6 @@
 ﻿using System.Net.Http;
 using System.Text;
 using System.Text.Json;
-using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Gridcoin.WebApi.Models;
 using Microsoft.Extensions.Logging;
@@ -50,49 +49,11 @@ namespace Gridcoin.WebApi.Services
             }
 
             var responseBody = await response.Content.ReadAsStringAsync();
-            var responseContent = JsonSerializer.Deserialize<T>(responseBody);
+            var responseContent = JsonSerializer.Deserialize<RpcResponse<T>>(responseBody, _jsonSerializerOptions);
 
             _logger.LogInformation($"{rpcRequest.Method} finished", responseContent);
 
-            return (true, responseContent);
+            return (true, responseContent.Result);
         }
-    }
-
-    public class Difficulty
-    {
-        public double Current { get; set; }
-        public double Target { get; set; }
-    }
-
-    public class GetInfoResponse
-    {        
-        public string Version { get; set; }
-
-        [JsonPropertyName("minor_version")]
-        public int MinorVersion { get; set; }
-
-        public int ProtocolVersion { get; set; }
-        public int XalletVersion { get; set; }
-        public double Balance { get; set; }
-        public double NewMint { get; set; }
-        public double Stake { get; set; }
-        public int Blocks { get; set; }
-
-        [JsonPropertyName("in_sync")]
-        public bool InSync { get; set; }
-
-        public int TimeOffset { get; set; }
-        public int UpTime { get; set; }
-        public double MoneySupply { get; set; }
-        public int Connections { get; set; }
-        public string Proxy { get; set; }
-        public string IP { get; set; }
-        public Difficulty Difficulty { get; set; }
-        public bool Testnet { get; set; }
-        public int KeyPoolOldest { get; set; }
-        public int KeyPoolSize { get; set; }
-        public double PayTXFee { get; set; }
-        public double MinInput { get; set; }
-        public string Errors { get; set; }
     }
 }
