+++ Gridcoin.WebApi/Startup.cs
