+++ Gridcoin.WebApi/appsettings.json
